telegram-phish/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.49
asyncpg==0.29.0
alembic==1.12.1
python-telegram-bot[job-queue]==20.6
jinja2==3.1.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.0.3